修改类型及添加模型

+++ demo-l/src/main/java/com/jianfei/d/dao/info/TemplateDao.java
@@ -12,6 +12,6 @@ import com.jianfei.d.entity.info.Template;
 
 public interface TemplateDao extends CrudDao<Template> {
 	
-	public List<Template> getListByType(Short type);
+	public List<Template> getListByType(Integer type);
 
 }
